refact: create randoms.js module
